+++ src/main/java/spring/boot/parser/service/RecordToDatabaseService.java
@@ -4,5 +4,5 @@ import com.univocity.parsers.common.record.Record;
 import java.util.List;
 
 public interface RecordToDatabaseService {
-    List<Record> getRecords();
+    void saveEntities(List<Record> records);
 }
