+++ src/main/java/spring/boot/parser/service/implementation/RecordToDatabaseServiceImpl.java
@@ -1,74 +1,107 @@
 package spring.boot.parser.service.implementation;
 
-import com.univocity.parsers.common.DataProcessingException;
-import com.univocity.parsers.common.ParsingContext;
-import com.univocity.parsers.common.RetryableErrorHandler;
-import com.univocity.parsers.common.processor.RowListProcessor;
 import com.univocity.parsers.common.record.Record;
-import com.univocity.parsers.csv.CsvParser;
-import com.univocity.parsers.csv.CsvParserSettings;
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.Arrays;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 import lombok.extern.log4j.Log4j;
-import org.springframework.beans.factory.annotation.Value;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import spring.boot.parser.model.Product;
+import spring.boot.parser.model.Review;
+import spring.boot.parser.model.Role;
+import spring.boot.parser.model.User;
+import spring.boot.parser.service.ProductService;
 import spring.boot.parser.service.RecordToDatabaseService;
+import spring.boot.parser.service.ReviewService;
+import spring.boot.parser.service.RoleService;
+import spring.boot.parser.service.UserService;
 
 @Log4j
 @Service
 public class RecordToDatabaseServiceImpl implements RecordToDatabaseService {
-    @Value("${input.file.path}")
-    private String fileName;
+    public static final String USER_ID = "UserId";
+    public static final String USER = "USER";
+    public static final String PRODUCT_ID = "ProductId";
+    public static final String PROFILE_NAME = "ProfileName";
+    public static final String ID = "Id";
+    public static final String HELPFULNESS_NUMERATOR = "HelpfulnessNumerator";
+    public static final String HELPFULNESS_DENOMINATOR = "HelpfulnessDenominator";
+    public static final String SCORE = "Score";
+    public static final String TIME = "Time";
+    public static final String SUMMARY = "Summary";
+    public static final String TEXT = "Text";
+    public static final String PASSWORD = "1111";
+    private final ReviewService reviewService;
+    private final UserService userService;
+    private final ProductService productService;
+    private final RoleService roleService;
 
-    @Override
-    public List<Record> getRecords() {
-        log.info("Starting to build parser at getRows() method");
-        CsvParser parser = new CsvParser(getCsvParserSettings());
-        try (BufferedReader bufferedReader = Files.newBufferedReader(Paths.get(fileName))) {
-            return parser.parseAllRecords(bufferedReader);
-        } catch (IOException e) {
-            throw new RuntimeException("File not found! " + fileName);
-        }
+    @Autowired
+    public RecordToDatabaseServiceImpl(ReviewService reviewService,
+                                       UserService userService,
+                                       ProductService productService,
+                                       RoleService roleService) {
+        this.reviewService = reviewService;
+        this.userService = userService;
+        this.productService = productService;
+        this.roleService = roleService;
     }
 
-    private CsvParserSettings getCsvParserSettings() {
-        CsvParserSettings parserSettings = new CsvParserSettings();
-        parserSettings.setLineSeparatorDetectionEnabled(true);
-        parserSettings.setNullValue("null");
-        parserSettings.setEmptyValue("");
-        parserSettings.setDelimiterDetectionEnabled(true);
-        parserSettings.setIgnoreLeadingWhitespaces(true);
-        parserSettings.setQuoteDetectionEnabled(true);
-        parserSettings.setSkipEmptyLines(true);
-        parserSettings.setHeaderExtractionEnabled(true);
-        parserSettings.setMaxCharsPerColumn(25000);
-        parserSettings.setIgnoreTrailingWhitespacesInQuotes(true);
-        parserSettings.setReadInputOnSeparateThread(true);
-        parserSettings.setErrorContentLength(10);
-        parserSettings.setAutoConfigurationEnabled(true);
-        parserSettings.setHeaderExtractionEnabled(true);
-        parserSettings.setProcessorErrorHandler(new RetryableErrorHandler<ParsingContext>() {
-            @Override
-            public void handleError(DataProcessingException error,
-                                    Object[] inputRow, ParsingContext context) {
-                System.out.println("Error processing row: " + Arrays.toString(inputRow));
-                System.out.println("Error details: column '" + error.getColumnName()
-                        + "' (index " + error.getColumnIndex() + ") has value '"
-                        + inputRow[error.getColumnIndex()] + "'. Setting it to null");
+    @Override
+    public void saveEntities(List<Record> records) {
+        log.info("Starting to parse List<Record>");
+        Set<User> userHashSet = new HashSet<>();
+        Set<Product> productHashSet = new HashSet<>();
 
-                if (error.getColumnIndex() == 0) {
-                    setDefaultValue(null);
-                } else {
-                    keepRecord();
-                }
-            }
-        });
-        RowListProcessor rowProcessor = new RowListProcessor();
-        parserSettings.setProcessor(rowProcessor);
-        return parserSettings;
+        for (Record row : records) {
+            userHashSet.add(parseUser(row));
+            productHashSet.add(parseProduct(row));
+        }
+
+        userService.saveAll(userHashSet);
+        log.info("Saving of the users are successful!");
+
+        productService.saveAll(productHashSet);
+        log.info("Saving of the products are successful!");
+
+        reviewService.saveAll(records.stream()
+                .map(this::parseReview)
+                .collect(Collectors.toList()));
+        log.info("Saving of the reviews are successful!");
+    }
+
+    private User parseUser(Record row) {
+        Set<Role> roleUserSet = new HashSet<>();
+        roleUserSet.add(roleService.getRoleByRoleName(USER));
+        return User.builder()
+                .userId(row.getString(USER_ID))
+                .profileName(row.getString(PROFILE_NAME))
+                .password(PASSWORD)
+                .role(roleUserSet)
+                .build();
+    }
+
+    private Product parseProduct(Record row) {
+        return Product.builder()
+                .productId(row.getString(PRODUCT_ID))
+                .build();
+    }
+
+    private Review parseReview(Record row) {
+        User user = userService.getByUserId(row.getString(USER_ID));
+        Product product = productService.getByProductId(row.getString(PRODUCT_ID));
+        return Review.builder()
+                .reviewId((row.getLong(ID)))
+                .helpfulnessNumerator(row.getLong(HELPFULNESS_NUMERATOR))
+                .helpfulnessDenominator(row.getLong(HELPFULNESS_DENOMINATOR))
+                .score(row.getLong(SCORE))
+                .time(row.getLong(TIME))
+                .summary(row.getString(SUMMARY).trim())
+                .text(row.getString(TEXT).trim())
+                .user(user)
+                .product(product)
+                .build();
     }
 }
