+++ src/main/resources/application.properties
@@ -1,8 +1,7 @@
 spring.h2.console.enabled=true
-
 spring.datasource.url=jdbc:h2:mem:testdb
 spring.datasource.driverClassName=org.h2.Driver
 spring.datasource.username=user
 spring.datasource.password=
 
-input.file.path=src/main/resources/test_file_ok.csv
+input.file.path=src/main/resources/reviews_file.csv
