+++ src/main/resources/test_file_ok.csv
@@ -1,0 +1,11 @@
+Id,ProductId,UserId,ProfileName,HelpfulnessNumerator,HelpfulnessDenominator,Score,Time,Summary,Text
+1965,B001FQ0UEE,ABWUZU0SCL5PW,Howard M.,2,2,5,1313539200,Absolutely wonderful,"A family favorite.  These are really unique treats.  The wonderful taste of flowers.  The violet and rose are great.  Orange blossom wasn't to my liking.  The company that makes them have been around for hundreds of years.  They are made in a small French town that was featured in a movie called ""Chocolat."""
+1620,B001RVFDOO,ATGKNPXJIBE42,Carol Sausen,0,0,4,1322006400,I like all but one,"I like all the varieties except the salt and pepper.  They taste too much like black pepper and nothing else.  The others are very goodl<a href=""http://www.amazon.com/gp/product/B001RVFDOO"">Popchips 6-Flavor Variety Pack, 0.8-Ounce Single Serve Bags (Pack of 24)</a>"
+4600,B0013ENGXK,A1B0ZKGS0ZAHQW,"V. Mcknelly ""weiserv""",0,0,5,1174003200,Use it every day,"I drink this tea every day (I'm 7 months pregnant), and have done so since I found out I was pregnant.  I love that it's organic and contains all the herbs that support a healthy pregnancy.  I don't notice anything dramatic when I drink it or don't drink it, but I will say I feel nourished when I drink it, and have had NO PREGNANCY COMPLICATIONS WHATSOEVER.  I'm 36, this is my first pregnancy, and I have been exceptionally healthy throughout.  I highly recommend this for any pregnant woman--at least a cup a day!"
+4277,B004391DK0,AJK1BSTTTO4F6,Alan W Corley,0,0,5,1290643200,Love me some gluten free bisquick!!!,"We just bought this the other day because for about 2 years we have found that biscuits are a very hard wheat/gluten free item to make...I LOVE IT!!! I even went further and put holes in the middle of the biscuit dough, flattened them a bit and then fried them in a bit of oil, sprinkled with sugar and had DOUGHNUTS! My 3 year old feels like he is getting a treat!"
+812,B0030C9A60,APNMGFU73W7ZC,Sharon Knibbe,0,0,5,1323388800,Scottie,The holiday liquor cups were exactly what I was looking for.  The chocolate is<br />delicious and holds up well to the liquors
+9131,B006N3IG4K,AHRASE2UFPFKK,Tom T,0,0,4,1330214400,Wolfgang's Hazlenut,"This one is my wife's favorite of all the Hazlenut blends she has tried. This is her ""go to"" after a busy day of being retired.<br />Delivery was exceptional as usual, could not ask for much more."
+7498,B0000T15GY,A1L8IWU5Y8TMEY,Maihsmy Boyle,0,0,5,1307577600,Awesome!,"I highly recommend these candies, they are that best! They are really good for the price. They come individually wraped which is nice, the colors are direrent from picture, they are bright colors which I love, because I can give it to girls and boys! They are also Oriental Traiding Brand and in their website, they have really good reviews for this item just look on their website an look for this item #5/923, I can not be more happy, the kids loves them!"
+8220,B000BXUVYG,AN0E0W3KTBPY,JST,1,1,5,1274832000,My Cats LOVE Them!!,"And they never vomit. They wolf them down, thought I think they are crunching. They haven't TOTALLY stopped their occasional hairball hacking, but it is rare. Perhaps they would completely solve it if I could AFFORD to give them to them all the time. Only problem with this product is $$$. I would love to buy it in a larger bag . . . if they just made one. If your cat's don't like 'em move on, but definately worth a try if you want a treat that also helps with hairballs."
+5179,B003FWKCB2,A137OMDVBAM3A8,fruitsnacklover,0,0,5,1313280000,Delish!,"I found these little awesomely delicious fruit snacks at a gluten free show in Wisconsin.The guy that was there offered me a package and I fell in love!They are organic, which makes them all the more cool, and they are very, Very tasty!I looked on the ingredients and was surprised to find no high fructose corn syrup!They're a little chewy at first, but I can not find anything else wrong with them. They are a wonderful fruit snack that I don't feel guilty about eating afterwards!"
+8413,B003VXFK44,A3GN7BRTTCMOMM,Kika,0,0,5,1331596800,Ahhhh...,This coffee is so smoothe.  I brew my morning cup with the medium cup size and it is delicious. It is a little weak on the large cup size setting.
